[功能] 后端-机房管理
关联 #[1729096900673543]后端-机房管理 http://192.168.0.96:8090/demo/rdm.html#/task-detail/939050947543040/939050947543050/1729096900673543

+++ src/main/java/neatlogic/module/cmdb/service/attr/AttrServiceImpl.java
@@ -43,6 +43,7 @@ import neatlogic.module.cmdb.dao.mapper.cientity.CiEntityMapper;
 import neatlogic.module.cmdb.dao.mapper.cischema.CiSchemaMapper;
 import neatlogic.module.cmdb.dao.mapper.transaction.TransactionMapper;
 import neatlogic.module.cmdb.fulltextindex.enums.CmdbFullTextIndexType;
+import neatlogic.module.cmdb.service.cientity.AttrInvokeManager;
 import neatlogic.module.cmdb.service.cientity.CiEntityService;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
@@ -74,6 +75,9 @@ public class AttrServiceImpl implements AttrService {
     @Resource
     private CiSchemaMapper ciSchemaMapper;
 
+    @Resource
+    private AttrInvokeManager attrInvokeManager;
+
 
     @Override
     @Transactional
@@ -87,7 +91,7 @@ public class AttrServiceImpl implements AttrService {
         attrVo.setCiVo(ciVo);
         IAttrValueHandler handler = AttrValueHandlerFactory.getHandler(attrVo.getType());
         handler.afterInsert(attrVo);
-        if (!handler.isNeedTargetCi()) {
+        if (!handler.isNeedTargetCi() && handler.isNeedCiEntityColumn()) {
             //由于以下操作是DDL操作，所以需要使用EscapeTransactionJob避开当前事务，否则在进行DDL操作之前事务就会提交，如果DDL出错，则上面的事务就无法回滚了
             CountDownLatch latch = new CountDownLatch(1);
             EscapeTransactionJob.State s = new EscapeTransactionJob(() -> {
@@ -164,7 +168,7 @@ public class AttrServiceImpl implements AttrService {
         attrVo.setCiVo(ciVo);
 
         handler.afterUpdate(attrVo);
-        if (!handler.isNeedTargetCi()) {
+        if (!handler.isNeedTargetCi() && handler.isNeedCiEntityColumn()) {
             //补充字典到全文检索
             if (Objects.equals(attrVo.getIsTerm(), 1)) {
                 IFullTextIndexHandler fullTextHandler = FullTextIndexHandlerFactory.getHandler(CmdbFullTextIndexType.CIENTITY.getType());
@@ -320,16 +324,20 @@ public class AttrServiceImpl implements AttrService {
             //删除引用属性数据
             ciEntityMapper.deleteAttrEntityByAttrId(attrVo.getId());
 
-            //删除模型属性
-            attrMapper.deleteAttrById(attrVo.getId());
-
             //某些类型的属性可能有删除后续操作
             IAttrValueHandler handler = AttrValueHandlerFactory.getHandler(attrVo.getType());
             handler.afterDelete(attrVo);
+
+            //删除外部存储属性的引用数据
+            attrInvokeManager.deleteByAttrId(attrVo.getId());
+
+            //删除模型属性
+            attrMapper.deleteAttrById(attrVo.getId());
             TransactionUtil.commitTx(tx);
             //物理删除字段
             //由于以上事务中的dml操作包含了以下ddl操作的表，如果使用 EscapeTransactionJob会导致事务等待产生死锁，所以这里不再使用EscapeTransactionJob去保证事务一致性。即使ddl删除字段失败，以上事务也会提交
-            if (!attrVo.isNeedTargetCi()) {
+            if (!attrVo.isNeedTargetCi()
+                    && ciSchemaMapper.checkColumnIsExists(TenantContext.get().getDataDbName(), attrVo.getCiId(), attrVo.getId()) > 0) {
                 ciSchemaMapper.deleteAttrFromCiTable(attrVo.getCiId(), attrCi.getCiTableName(), attrVo);
             }
         } catch (Exception ex) {
